Unify DEBUG behavior, minor cleaning
All of the debug #ifdefs use the DEBUG define. This will
generate much larger debug log files, at least until I have
time to do something about that. There are a couple of places
where I have just commented out the debug message because it
would not compile, and there is still a DEBUGXXX in there to
mark a non-IO debug action.